Single quote

+++ client/.eslintrc.js
@@ -14,7 +14,7 @@ module.exports = {
     "angular/di": ['error', 'array'],
     "indent": ['error', 2],
     "linebreak-style": ['off'],
-    "quotes": ['error', 'double', { avoidEscape: false, allowTemplateLiterals: true }],
+    "quotes": ['error', 'single', { avoidEscape: false, allowTemplateLiterals: true }],
     "semi": ['error', 'always'],
     "no-var": 'off', // We don't use ES6
     "no-param-reassign": ['error', { props: false }],
